+++ municipal-services/fsm/src/main/java/org/egov/fsm/web/model/PitDetail.java
@@ -13,6 +13,9 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import javax.validation.constraints.DecimalMax;
+import javax.validation.constraints.Max;
+
 /**
  * cature the pit details 
  */
@@ -41,14 +44,18 @@ public class PitDetail   {
   private String tenantId = null;
   
   @JsonProperty("height")
+  @DecimalMax(value="99.9")
   private Double height = null;
 
   @JsonProperty("length")
+  @DecimalMax(value="99.9",message="Length value sholud be less than 100")
   private Double length = null;
 
   @JsonProperty("width")
+  @DecimalMax(value="99.9",message="Width value sholud be less than 100")
   private Double width = null;
-  
+ 
+  @DecimalMax(value="99.9", message="Diameter value sholud be less than 100")
   @JsonProperty("diameter")
   private Double diameter = null;