+++ municipal-services/fsm/src/main/java/org/egov/fsm/web/model/user/User.java
@@ -44,6 +44,7 @@ public class User {
 
     @Size(min=1,max=64)
     @SafeHtml
+    @Pattern(regexp = "^[a-zA-Z \\-'`\\.]*$", message = "Invalid name. Only alphabets and special characters -, ',`, .")
     @JsonProperty("name")
     private String name;