+++ municipal-services/property-services/src/main/java/org/egov/pt/config/PropertyConfiguration.java
@@ -316,4 +316,34 @@ public class PropertyConfiguration {
 
     @Value("${state.level.tenant.id}")
     private String stateLevelTenantId;
+
+
+    // ES Config
+
+    @Value("${elasticsearch.host}")
+    private String esHost;
+
+    @Value("${property.es.index}")
+    private String esPTIndex;
+
+    @Value("${elasticsearch.search.endpoint}")
+    private String esSearchEndpoint;
+
+    @Value("${pt.search.name.fuziness}")
+    private String nameFuziness;
+
+    @Value("${pt.search.doorno.fuziness}")
+    private String doorNoFuziness;
+
+    @Value("${pt.search.oldpropertyid.fuziness}")
+    private String oldPropertyIdFuziness;
+
+    @Value("${pt.fuzzy.searh.is.wildcard}")
+    private Boolean isSearchWildcardBased;
+
+    // Inbox service Config
+
+    @Value("${inbox.property.search.allowed}")
+    private Boolean isInboxSearchAllowed;
+
 }