+++ municipal-services/tl-services/src/main/java/org/egov/tl/service/TradeLicenseService.java
@@ -168,7 +168,7 @@ public class TradeLicenseService {
         tlValidator.validateSearch(requestInfo,criteria,serviceFromPath, isInterServiceCall);
         criteria.setBusinessService(serviceFromPath);
         enrichmentService.enrichSearchCriteriaWithAccountId(requestInfo,criteria);
-         if(criteria.getMobileNumber()!=null){
+         if(criteria.getMobileNumber()!=null || criteria.getOwnerName() != null){
              licenses = getLicensesFromMobileNumber(criteria,requestInfo);
          }
          else {