+++ municipal-services/vehicle/src/main/java/org/egov/vehicle/trip/repository/VehicleTripRepository.java
@@ -3,6 +3,8 @@ package org.egov.vehicle.trip.repository;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.egov.common.contract.request.RequestInfo;
 import org.egov.tracer.model.CustomException;
 import org.egov.vehicle.config.VehicleConfiguration;
@@ -17,9 +19,13 @@ import org.egov.vehicle.trip.web.model.VehicleTripResponse;
 import org.egov.vehicle.trip.web.model.VehicleTripSearchCriteria;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.SingleColumnRowMapper;
 import org.springframework.stereotype.Repository;
 
+import lombok.extern.slf4j.Slf4j;
+
 @Repository
+@Slf4j
 public class VehicleTripRepository {
 	
 	@Autowired
@@ -107,6 +113,31 @@ public class VehicleTripRepository {
 		}
 		
 		return ids;
+	}
+
+	public List<String> fetchVehicleTripIds(@Valid VehicleTripSearchCriteria criteria) {
+
+		List<Object> preparedStmtList = new ArrayList<>();
+		preparedStmtList.add(criteria.getOffset());
+		preparedStmtList.add(criteria.getLimit());
+
+		List<String> ids = jdbcTemplate.query("SELECT id from eg_vehicle_trip ORDER BY createdtime offset " +
+						" ? " +
+						"limit ? ",
+				preparedStmtList.toArray(),
+				new SingleColumnRowMapper<>(String.class));
+		return ids;
+	}
+
+	public List<VehicleTrip> getVehicleTripPlainSearch(VehicleTripSearchCriteria criteria) {
+		if(criteria.getIds() == null || criteria.getIds().isEmpty())
+			throw new CustomException("PLAIN_SEARCH_ERROR", "Search only allowed by ids!");
+
+		List<Object> preparedStmtList = new ArrayList<>();
+		String query = queryBuilder.getvehicleTripLikeQuery(criteria, preparedStmtList);
+		log.info("Query: "+query);
+		log.info("PS: "+preparedStmtList);
+		return jdbcTemplate.query(query, preparedStmtList.toArray(), mapper);
 	}
 
 }