+++ municipal-services/fsm/CHANGELOG.md
@@ -2,6 +2,13 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.0.3
+
+- Added condition to check the pit dimention to be less than 100.
+- Make Property and sub property as mandetory while creating FSM.
+- FSTP Plant mapping changes.
+- Added plain search service.
+
 ## 1.0.2
 
 - Fixed security issue of untrusted data pass as user input.